requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-admin-tools==0.5.2
django-exclusivebooleanfield==0.3.0
django-grappelli==2.6.4
-e git+https://github.com/FundingGates/django-http-proxy.git@8e2371b2bae76b564e2bd47321471aa2a418b773#egg=django_http_proxy
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.4.5.1
docopt==0.4.0
gunicorn==19.3.0
html2text==2015.6.12
mailchimp==2.0.9
psycopg2==2.6
requests==2.6.0
six==1.9.0
static3==0.5.1
vk==1.5.3
